# backend/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Raptr Automation API",
    "description": "Automation + Manual Test management API",
    "version": "1.0.0"
  },
  "host": "localhost:6223",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/cases/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "folderId",
            "in": "query",
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "folderId",
            "in": "query",
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                },
                "state": {
                  "example": "any"
                },
                "priority": {
                  "example": "any"
                },
                "type": {
                  "example": "any"
                },
                "automationStatus": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "template": {
                  "example": "any"
                },
                "preConditions": {
                  "example": "any"
                },
                "expectedResults": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/cases/{caseId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "caseId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "caseId",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "caseId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "caseId",
            "in": "query",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "Steps": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/runcases/executions/testcase/{testcase_id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "testcase_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/autotestcases/project/{projectId}/modules/{module}/cases": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "module",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/autotestcases/": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "module": {
                  "example": "any"
                },
                "testfilename": {
                  "example": "any"
                },
                "testdata": {
                  "example": "any"
                },
                "status": {
                  "example": "any"
                },
                "created_by": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/autotestcases/{ts_id}": {
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "ts_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "ts_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/autotestcasegroups/": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "group_name": {
                  "example": "any"
                },
                "ts_ids": {
                  "example": "any"
                },
                "testdataPath": {
                  "example": "any"
                },
                "created_by": {
                  "example": "any"
                },
                "status": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/autotestcasegroups/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "ts_ids": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/autotestcasegroups/{id}/cases": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/bugzilla/bugs/testcase": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "testcaseIds",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/testcases/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "type": "string"
          },
          {
            "name": "projectId",
            "in": "query",
            "type": "string"
          },
          {
            "name": "ts_sid",
            "in": "query",
            "type": "string"
          },
          {
            "name": "ts_id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "tc_id",
            "in": "query",
            "type": "string"
          },
          {
            "name": "tc_type",
            "in": "query",
            "type": "string"
          },
          {
            "name": "status",
            "in": "query",
            "type": "string"
          },
          {
            "name": "spoc",
            "in": "query",
            "type": "string"
          },
          {
            "name": "createdFrom",
            "in": "query",
            "type": "string"
          },
          {
            "name": "createdTo",
            "in": "query",
            "type": "string"
          },
          {
            "name": "module",
            "in": "query",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          },
          {
            "name": "offset",
            "in": "query",
            "type": "string"
          },
          {
            "name": "sortBy",
            "in": "query",
            "type": "string"
          },
          {
            "name": "sortDir",
            "in": "query",
            "type": "string"
          },
          {
            "name": "includeDeleted",
            "in": "query",
            "type": "string"
          },
          {
            "name": "includeScenario",
            "in": "query",
            "type": "string"
          },
          {
            "name": "buildId",
            "in": "query",
            "type": "string"
          },
          {
            "name": "intent",
            "in": "query",
            "type": "string"
          },
          {
            "name": "priority",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "projectId": {
                  "example": "any"
                },
                "ts_sid": {
                  "example": "any"
                },
                "ts_id": {
                  "example": "any"
                },
                "tc_id": {
                  "example": "any"
                },
                "tc_description": {
                  "example": "any"
                },
                "tc_type": {
                  "example": "any"
                },
                "prerequisites": {
                  "example": "any"
                },
                "steps": {
                  "example": "any"
                },
                "tc_data": {
                  "example": "any"
                },
                "expected_results": {
                  "example": "any"
                },
                "actual_result": {
                  "example": "any"
                },
                "spoc": {
                  "example": "any"
                },
                "status": {
                  "example": "any"
                },
                "pot_link": {
                  "example": "any"
                },
                "bug_id": {
                  "example": "any"
                },
                "Iteration": {
                  "example": "any"
                },
                "remarks": {
                  "example": "any"
                },
                "count": {
                  "example": "any"
                },
                "tracability": {
                  "example": "any"
                },
                "testers": {
                  "example": "any"
                },
                "priority": {
                  "example": "any"
                },
                "intent": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/testcases/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "includeScenario",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "projectId": {
                  "example": "any"
                },
                "ts_sid": {
                  "example": "any"
                },
                "ts_id": {
                  "example": "any"
                },
                "tc_id": {
                  "example": "any"
                },
                "tc_description": {
                  "example": "any"
                },
                "tc_type": {
                  "example": "any"
                },
                "prerequisites": {
                  "example": "any"
                },
                "steps": {
                  "example": "any"
                },
                "rawSteps": {
                  "example": "any"
                },
                "tc_data": {
                  "example": "any"
                },
                "expected_results": {
                  "example": "any"
                },
                "actual_result": {
                  "example": "any"
                },
                "spoc": {
                  "example": "any"
                },
                "status": {
                  "example": "any"
                },
                "pot_link": {
                  "example": "any"
                },
                "bug_id": {
                  "example": "any"
                },
                "Iteration": {
                  "example": "any"
                },
                "remarks": {
                  "example": "any"
                },
                "count": {
                  "example": "any"
                },
                "tracability": {
                  "example": "any"
                },
                "testers": {
                  "example": "any"
                },
                "priority": {
                  "example": "any"
                },
                "intent": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/{id}/cases": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/project/{projectId}/modules/{module}/cases": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "module",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/testcase": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "testcaseIds",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/testcase/{testcase_id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "testcase_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}